\name{GenomicRanges-deprecated}
\alias{GenomicRanges-deprecated}
\title{Deprecated Functions in package \pkg{GenomicRanges}}

\alias{makeSeqnameIds}

\description{
  The functions or variables listed here have been deprecated and should no
  longer be used.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{makeSeqnameIds: \code{\link[GenomeInfoDb]{rankSeqlevels}}}
  }
}

\seealso{
  \code{\link{Deprecated}}
}
