\name{makeSummarizedExperimentFromExpressionSet}

\alias{makeSummarizedExperimentFromExpressionSet}

\title{Make a SummarizedExperiment object from an ExpressionSet}

\description{
  WARNING: The SummarizedExperiment class is deprecated and being
  replaced with the \link[SummarizedExperiment]{RangedSummarizedExperiment}
  class defined in the new \pkg{SummarizedExperiment} package.

  Please make sure to install and load the \pkg{SummarizedExperiment} package
  where the \code{makeSummarizedExperimentFromExpressionSet} function is now
  defined and documented. Note that the function now returns a
  \link[SummarizedExperiment]{RangedSummarizedExperiment} instance instead
  of a \link{SummarizedExperiment} instance.
}

