\name{GenomicRanges-deprecated}
\alias{GenomicRanges-deprecated}
\title{Deprecated and defunct functions in package \pkg{GenomicRanges}}

\alias{makeSeqnameIds}

\description{
  The functions or variables listed here have been deprecated or defunct
  and should no longer be used.
}

\details{
  The following function is defunct; use the replacement indicated below:
  \itemize{

    \item{makeSeqnameIds: \code{\link[GenomeInfoDb]{rankSeqlevels}}}
  }
}

\seealso{
  \code{\link{Deprecated}} and \code{\link{Defunct}}
}
